chore(tests): extract progress-callback tests — test_engine.py under the 800-LOC ceiling
CodeRabbit finding on #500: the first extraction (938->867) still left
tests/test_engine.py over the module limit. TestProgressCallbacks (91
lines) moves to tests/test_engine_progress_callbacks.py; content
unchanged; test_engine.py now 776 LOC.
